requirements.txt
fastapi
uvicorn
rank-bm25
sentence-transformers
faiss-cpu
numpy
pandas
streamlit
requests
pytest
pydantic
sqlite-utils